Serial.begin parameter is long, not int

+++ include/arduino-mock/serial.h
@@ -74,7 +74,7 @@ class Serial_ {
     size_t write(uint8_t);
     size_t write(const char *str);
 
-    uint8_t begin(uint16_t);
+    uint8_t begin(uint32_t);
 
     uint8_t available();
     uint8_t read();
